+++ client/src/components/auth/SingUp/index.js
@@ -3,3 +3,5 @@ export { default as FormInput } from './FormInput';
 export { default as PasswordInput } from './PasswordInput';
 export { default as BirthDateInput } from './BirthDateInput';
 export { default as Input } from './Input';
+export { default as SignUpForm } from './SignUpForm';
+export { default as SignUp } from './SignUp';
